Primitivization and dual state of transaction state collections
Switch transaction state collections to be primitive.
Introduce primitive set and map that allows dual state: active and stable.
Active - actively changing state (actual state).
Stable - state of the collection on some particular moment.

+++ community/kernel/src/main/java/org/neo4j/kernel/api/properties/PropertyKeyValue.java
@@ -73,7 +73,7 @@ public class PropertyKeyValue implements StorageProperty
     @Override
     public int hashCode()
     {
-        int result = propertyKeyId;
+        int result = (int) propertyKeyId;
         result = 31 * result + value.hashCode();
         return result;
     }
